Adds SerilogServiceCollectionExtensions
Allows for more flexibility when configuring serilog services during startup. In some cases we don't want to use the serilog static Log and we aren't able to pre-create an ILogger before the HostBuilder is configured. We will create our own serilog logger during Configure Services and then we need to configure the rest of serilog services which isn't possible without copying all this code it would be nicer to just expose these methods so people can use them.